+++ ChangeLog.md
@@ -51,7 +51,11 @@ Different versions of mealpy in terms of passing hyper-parameters. So please car
 + Add `Owl Search Algorithm (OSA)` to the `swarm-based` group
 + Add `Dandelion Optimizer (DandelionO)` to the `swarm-based` group
 + Add `Red Fox Optimization (RFO)` to the `swarm-based` group
-
++ Add `Crayfish Optimization Algorithm (CrayfishOA)` to the `swarm-based` group
++ Add `Red Fox Optimization (RFO)` to the `swarm-based` group
++ Add `Red Fox Optimization (RFO)` to the `swarm-based` group
++ Add `Red Fox Optimization (RFO)` to the `swarm-based` group
++ Add `Red Fox Optimization (RFO)` to the `swarm-based` group
 
 ---